Completion of the Account classes
Completed the migration and refactoring of the User class to Account,
and the relevant service and repository classes. As such the User class
has been removed in favor of the Account class.

Updated the fields in the Itinerary classes to more accurately reflect
the data that will be stored in the database and to be more realistic
in a sense.

Authentication is not fully complete for the ItineraryController, but
you can now create, read and delete an itinerary for a user.
You can't update entries but will be implemented soon.

Still no tests, but will be implemented soon.

+++ src/main/java/com/beastlymc/triptimize/itinerary/ItineraryService.java
@@ -6,16 +6,19 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import lombok.AllArgsConstructor;
+import com.beastlymc.triptimize.user.AccountRepository;
+
+import lombok.RequiredArgsConstructor;
 
 /**
  * Service for {@link Itinerary} objects.
  */
-@AllArgsConstructor
+@RequiredArgsConstructor
 @Service
 public class ItineraryService {
     
     private final ItineraryRepository itineraryRepository;
+    private final AccountRepository accountRepository;
 
     /**
      * Finds an itinerary by its id.
@@ -23,11 +26,13 @@ public class ItineraryService {
      * @param id the id to search for
      * @return an Optional containing the itinerary if found, or an empty Optional otherwise
      */
-    public ResponseEntity<Itinerary> findById(Long id) {
+    public ResponseEntity<Itinerary> findById(final Integer id) {
         Optional<Itinerary> itinerary = itineraryRepository.findById(id);
+        // Check if the itinerary exists
         if (itinerary.isPresent()) {
             return ResponseEntity.ok(itinerary.get());
         }
+        
         return ResponseEntity.notFound().build();
     }
 
@@ -46,17 +51,26 @@ public class ItineraryService {
      * @param itinerary the itinerary to save
      * @return the saved itinerary
      */
-    public ResponseEntity<Itinerary> save(ItineraryRequest request) {
-        Itinerary itinerary = new Itinerary();
+    public ResponseEntity<Itinerary> saveItinerary(final ItineraryRequest request) {
+        // Check if the owner exists
+        if(accountRepository.existsById(request.getOwnerId()) == false) {
+            return ResponseEntity.badRequest().build();
+        }
 
-        itinerary.setName(request.getName());
-        itinerary.setDescription(request.getDescription());
-        itinerary.setStartDate(request.getStartDate());
-        itinerary.setEndDate(request.getEndDate());
-        itinerary.setActivities(request.getActivities());
-        itinerary.setPublic(true);
+        Itinerary itinerary = Itinerary.builder()
+        .author(accountRepository.getReferenceById(request.getOwnerId()))
+        .name(request.getName())
+        .location(request.getLocation())
+        .description(request.getDescription())
+        .startDate(request.getStartDate())
+        .endDate(request.getEndDate())
+        .activities(request.getActivities())
+        .isPublic(request.isPublic())
+        .build();
 
-        return ResponseEntity.ok(itineraryRepository.save(itinerary));
+        itineraryRepository.save(itinerary);
+
+        return ResponseEntity.ok(itinerary);
     }
 
     /**
@@ -64,10 +78,48 @@ public class ItineraryService {
      *
      * @param id the id of the itinerary to delete
      */
-    public ResponseEntity<Long> deleteById(Long id) {
+    public ResponseEntity<Integer> deleteById(final Integer id) {
+        // Check if the itinerary exists
+        if(!itineraryRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
+
         itineraryRepository.deleteById(id);
         return ResponseEntity.ok(id);
     }
 
+    /**
+     * Updates an itinerary.
+     *
+     * @param itinerary the itinerary to update
+     * @return the updated itinerary
+     */
+    public ResponseEntity<Itinerary> updateItinerary(final Integer id, final ItineraryRequest request) {
+        Optional<Itinerary> updatedItinerary = itineraryRepository.findById(id);
+
+        // Check if the itinerary exists
+        if (!updatedItinerary.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        // Check if the owner exists
+        if(accountRepository.existsById(request.getOwnerId()) == false) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        Itinerary updated = updatedItinerary.get();
+        updated.setName(request.getName());
+        updated.setAuthor(accountRepository.getReferenceById(request.getOwnerId()));
+        updated.setDescription(request.getDescription());
+        updated.setLocation(request.getLocation());
+        updated.setStartDate(request.getStartDate());
+        updated.setEndDate(request.getEndDate());
+        updated.setActivities(request.getActivities());
+        updated.setPublic(request.isPublic());
+
+        itineraryRepository.saveAndFlush(updated);
+
+        return ResponseEntity.ok(updated);
+    }
     
 }
